+++ app/Views/partials/footer.php
@@ -28,5 +28,6 @@
     const tooltipTriggerList = document.querySelectorAll('[data-bs-toggle="tooltip"]')
     const tooltipList = [...tooltipTriggerList].map(tooltipTriggerEl => new bootstrap.Tooltip(tooltipTriggerEl))
 </script>
+<script src="/public/js/teacher-profiles.js"></script>
 </body>
 </html>